refactor: 유틸함수 & 타입 분리 #96 - utils/getEmptyProfile - Payload Type
