+++ src/types/index.d.ts
@@ -8,6 +8,24 @@ type CommonResponse<T = any> = {
   data: T;
 };
 
+type OverlayProps = {
+  isOpen: boolean;
+  onClose: VoidFunction;
+};
+
+type Payload = {
+  authority: string;
+  exp: number;
+  id: string;
+};
+
+type FileValue = {
+  files: {
+    uuid: string;
+    path: string;
+  }[];
+};
+
 type GenderType = 'MALE' | 'FEMALE';
 
 type BodyType = '마름' | '탄탄 슬림' | '보통' | '통통' | '근육' | '탄탄_슬림';
